+++ src/asp/Tools/jitter_solve.cc
@@ -313,10 +313,11 @@ void handle_arguments(int argc, char *argv[], Options& opt, rig::RigSet & rig) {
      "Weight image for anchor points. Limits where anchor points are placed and their "
      "weight. These weights are additionally multiplied by --anchor-weight. See also "
      "--weight-image.")
-     ("smoothness-weight", po::value(&opt.smoothness_weight)->default_value(0.0),
-     "A weight to penalize the curvature of the sequence of orientations in the linescan "
-     "cameras being optimized being much larger than of the initial cameras. Increasing "
-     "this will make this sequence more smooth")
+     ("smoothness-weight", po::value(&opt.smoothness_weight)->default_value(1.0),
+      "A weight to penalize high-frequency changes in the sequence of orientations "
+      "in the linescan cameras being optimized. This is internally adjusted based "
+      "on the initial curvature of the sequence of orientations, and the default "
+      "should be good enough. A higher value may impede convergence.")
     ("initial-camera-constraint", 
      po::bool_switch(&opt.initial_camera_constraint)->default_value(false),
      "When constraining roll and yaw, measure these not in the satellite along-track/ "